refactor: Use single hooks
