+++ components/_util/hooks/useFlexGapSupport.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { detectFlexGapSupported } from '../../_util/styleChecker';
+import { detectFlexGapSupported } from '../styleChecker';
 
 export default () => {
   const [flexible, setFlexible] = React.useState(false);
